# server/config/swagger/Api.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: portal-scheduling
  description: Educational project
basePath: /api/v1/scheduling
...
consumes:
  - application/json
produces:
  - application/json
...
paths:
...
  /schedule-itens/{id}:
    get:
      tags:
        - ScheduleItens
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the schedule item to retrieve
          type: string
      responses:
        200:
          description: Return the schedule item with the specified id
        404:
          description: Schedule item not found
    delete:
      tags:
        - ScheduleItens
      description: Delete a existent schedule item
      parameters:
        - name: id
          in: path
          required: true
          description: The id of the schedule item to delete
          type: string
      responses:
        200:
          description: Return the schedule item that was deleted
        404:
          description: Schedule item not found
...